Move RowIterable to its own class
