+++ presto-main/src/main/java/com/facebook/presto/server/protocol/Query.java
@@ -34,17 +34,14 @@ import com.facebook.presto.execution.buffer.SerializedPage;
 import com.facebook.presto.metadata.SessionPropertyManager;
 import com.facebook.presto.operator.ExchangeClient;
 import com.facebook.presto.server.SessionContext;
-import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.ErrorCode;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.QueryId;
-import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockEncodingSerde;
 import com.facebook.presto.spi.type.StandardTypes;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeSignature;
 import com.facebook.presto.transaction.TransactionId;
-import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
@@ -62,8 +59,6 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 
 import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
@@ -591,59 +586,5 @@ class Query
                 errorCode.getType().toString(),
                 failure.getErrorLocation(),
                 failure);
-    }
-
-    private static class RowIterable
-            implements Iterable<List<Object>>
-    {
-        private final ConnectorSession session;
-        private final List<Type> types;
-        private final Page page;
-
-        private RowIterable(ConnectorSession session, List<Type> types, Page page)
-        {
-            this.session = session;
-            this.types = ImmutableList.copyOf(requireNonNull(types, "types is null"));
-            this.page = requireNonNull(page, "page is null");
-        }
-
-        @Override
-        public Iterator<List<Object>> iterator()
-        {
-            return new RowIterator(session, types, page);
-        }
-    }
-
-    private static class RowIterator
-            extends AbstractIterator<List<Object>>
-    {
-        private final ConnectorSession session;
-        private final List<Type> types;
-        private final Page page;
-        private int position = -1;
-
-        private RowIterator(ConnectorSession session, List<Type> types, Page page)
-        {
-            this.session = session;
-            this.types = types;
-            this.page = page;
-        }
-
-        @Override
-        protected List<Object> computeNext()
-        {
-            position++;
-            if (position >= page.getPositionCount()) {
-                return endOfData();
-            }
-
-            List<Object> values = new ArrayList<>(page.getChannelCount());
-            for (int channel = 0; channel < page.getChannelCount(); channel++) {
-                Type type = types.get(channel);
-                Block block = page.getBlock(channel);
-                values.add(type.getObjectValue(session, block, position));
-            }
-            return Collections.unmodifiableList(values);
-        }
     }
 }
